ci: split extensive vector search tests into another file
